tar ut janei radiopanel til eget spm komponent
