+++ src/components/JaNeiSpørsmål.tsx
@@ -1,0 +1,51 @@
+import React, { SyntheticEvent } from 'react';
+import { ISpørsmål, ISvar } from '../models/spørsmal';
+import { Element } from 'nav-frontend-typografi';
+import { hentSvar, hentTekstidTilJaNeiSvar } from '../utils/spørsmålogsvar';
+import { RadioPanel } from 'nav-frontend-skjema';
+import { injectIntl, IntlShape } from 'react-intl';
+import useSøknadContext from '../context/SøknadContext';
+
+interface Props {
+  spørsmål: ISpørsmål;
+  tekstid: string;
+  intl: IntlShape;
+}
+const JaNeiSpørsmål: React.FC<Props> = ({ spørsmål, tekstid, intl }) => {
+  const { søknad, settSøknad } = useSøknadContext();
+
+  const onClickHandle = (
+    e: SyntheticEvent<EventTarget, Event>,
+    spørsmål: ISpørsmål,
+    svar: ISvar
+  ): void => {
+    settSøknad({ ...søknad, [spørsmål.spørsmål_id]: svar === ISvar.JA });
+  };
+
+  return (
+    <div key={spørsmål.spørsmål_id} className="spørsmålgruppe">
+      <Element>{intl.formatMessage({ id: tekstid })}</Element>
+      <div className={'radioknapp__wrapper'}>
+        {spørsmål.svaralternativer.map((svar: ISvar) => {
+          const svarISøknad = hentSvar(spørsmål, svar, søknad);
+          return (
+            <div key={svar} className={'radioknapp__item'}>
+              <RadioPanel
+                key={svar}
+                name={spørsmål.spørsmål_id + svar}
+                label={intl.formatMessage({
+                  id: hentTekstidTilJaNeiSvar(svar),
+                })}
+                value={svar}
+                checked={svarISøknad ? svarISøknad : false}
+                onChange={(e) => onClickHandle(e, spørsmål, svar)}
+              />
+            </div>
+          );
+        })}
+      </div>
+    </div>
+  );
+};
+
+export default injectIntl(JaNeiSpørsmål);
