+++ src/config/SivilstatusConfig.tsx
@@ -1,7 +1,7 @@
 import { ISpørsmål, ISvar } from '../models/spørsmal';
 
 export const SpørsmålOgSvar: ISpørsmål = {
-  spørsmål_id: 'søkerOppholderSegINorge',
-  tekstid: 'medlemskap.spm.opphold',
+  spørsmål_id: 'søkerHarSøktSeparasjon',
+  tekstid: 'sivilstatus.søktseparasjon',
   svaralternativer: [ISvar.JA, ISvar.NEI],
 };
