+++ src/søknad/Søknad.tsx
@@ -5,7 +5,7 @@ import { Panel } from 'nav-frontend-paneler';
 import Medlemskap from './inngangsvilkår/personopplysninger/Medlemskap';
 import Personopplysninger from './inngangsvilkår/personopplysninger/Personopplysninger';
 import useSøknadContext from '../context/SøknadContext';
-import Sivilstatus from './inngangsvilkår/personopplysninger/Sivilstatus';
+import Sivilstatus from './inngangsvilkår/personopplysninger/sivilstatus/Sivilstatus';
 
 interface IState {
   status: string;
